requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
pymongo==4.5.0
werkzeug==2.3.7
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.0
opencv-python==4.8.0.76
